feat(EditProject): ✨ implemented new layouts
